Now can push --submit changes repeatedly without having to do a pull. Push now syncs the p4 workarea to the last changelist submitted from this client instead of the last changelist pulled.
Changed push so it reverts a newly created changelist if p4 raised
an error when creating it.

Changed p4pending to output all pending and submitted changelists
since the last pull, not just those that were pushed from Mercurial.
The first column now always shows the changelist number, and the submitted
or pending state is now shown in the second column of output.